+++ server/controller/http/router/vtap_group_config.go
@@ -22,9 +22,9 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/gin-gonic/gin/binding"
 
+	"github.com/deepflowio/deepflow/server/agent_config"
 	. "github.com/deepflowio/deepflow/server/controller/http/router/common"
 	"github.com/deepflowio/deepflow/server/controller/http/service"
-	"github.com/deepflowio/deepflow/server/controller/model"
 )
 
 type VTapGroupConfig struct{}
@@ -50,7 +50,7 @@ func (cgc *VTapGroupConfig) RegisterTo(e *gin.Engine) {
 }
 
 func createVTapGroupConfig(c *gin.Context) {
-	vTapGroupConfig := &model.VTapGroupConfiguration{}
+	vTapGroupConfig := &agent_config.VTapGroupConfiguration{}
 	err := c.ShouldBindBodyWith(&vTapGroupConfig, binding.JSON)
 	if err == nil {
 		data, err := service.CreateVTapGroupConfig(vTapGroupConfig)
@@ -68,7 +68,7 @@ func deleteVTapGroupConfig(c *gin.Context) {
 
 func updateVTapGroupConfig(c *gin.Context) {
 	lcuuid := c.Param("lcuuid")
-	vTapGroupConfig := &model.VTapGroupConfiguration{}
+	vTapGroupConfig := &agent_config.VTapGroupConfiguration{}
 	err := c.ShouldBindBodyWith(&vTapGroupConfig, binding.JSON)
 	if err == nil {
 		data, err := service.UpdateVTapGroupConfig(lcuuid, vTapGroupConfig)
@@ -101,7 +101,7 @@ func getVTapGroupAdvancedConfig(c *gin.Context) {
 
 func updateVTapGroupAdvancedConfig(c *gin.Context) {
 	lcuuid := c.Param("lcuuid")
-	vTapGroupConfig := &model.VTapGroupConfiguration{}
+	vTapGroupConfig := &agent_config.VTapGroupConfiguration{}
 	err := c.ShouldBindBodyWith(&vTapGroupConfig, binding.YAML)
 	if err == nil || err == io.EOF {
 		data, err := service.UpdateVTapGroupAdvancedConfig(lcuuid, vTapGroupConfig)
@@ -112,7 +112,7 @@ func updateVTapGroupAdvancedConfig(c *gin.Context) {
 }
 
 func createVTapGroupAdvancedConfig(c *gin.Context) {
-	vTapGroupConfig := &model.VTapGroupConfiguration{}
+	vTapGroupConfig := &agent_config.VTapGroupConfiguration{}
 	err := c.ShouldBindBodyWith(&vTapGroupConfig, binding.YAML)
 	if err == nil {
 		data, err := service.CreateVTapGroupAdvancedConfig(vTapGroupConfig)
