+++ server/controller/model/model.go
@@ -18,6 +18,8 @@ package model
 
 import (
 	"time"
+
+	"github.com/deepflowio/deepflow/server/agent_config"
 )
 
 type ControllerUpdate struct {
@@ -471,174 +473,9 @@ type AdditionalResource struct {
 	PeerConnections []AdditionalResourcePeerConnection `json:"peer_connections" yaml:"peer_connections,omitempty" binding:"omitempty,dive"`
 }
 
-type VTapGroupConfiguration struct {
-	VTapGroupID                       *string       `json:"VTAP_GROUP_ID" yaml:"vtap_group_id,omitempty"`
-	VTapGroupLcuuid                   *string       `json:"VTAP_GROUP_LCUUID" yaml:"vtap_group_lcuuid,omitempty"`
-	MaxCollectPps                     *int          `json:"MAX_COLLECT_PPS" yaml:"max_collect_pps,omitempty"`
-	MaxNpbBps                         *int64        `json:"MAX_NPB_BPS" yaml:"max_npb_bps,omitempty"` // unit: bps
-	MaxCPUs                           *int          `json:"MAX_CPUS" yaml:"max_cpus,omitempty"`
-	MaxMemory                         *int          `json:"MAX_MEMORY" yaml:"max_memory,omitempty"` // unit: M
-	SyncInterval                      *int          `json:"SYNC_INTERVAL" yaml:"sync_interval,omitempty"`
-	PlatformSyncInterval              *int          `json:"PLATFORM_SYNC_INTERVAL" yaml:"platform_sync_interval,omitempty"`
-	StatsInterval                     *int          `json:"STATS_INTERVAL" yaml:"stats_interval,omitempty"`
-	RsyslogEnabled                    *int          `json:"RSYSLOG_ENABLED" yaml:"rsyslog_enabled,omitempty"` // 0: disabled 1:enabled
-	SystemLoadCircuitBreakerThreshold *float32      `json:"SYSTEM_LOAD_CIRCUIT_BREAKER_THRESHOLD" yaml:"system_load_circuit_breaker_threshold,omitempty"`
-	SystemLoadCircuitBreakerRecover   *float32      `json:"SYSTEM_LOAD_CIRCUIT_BREAKER_RECOVER" yaml:"system_load_circuit_breaker_recover,omitempty"`
-	SystemLoadCircuitBreakerMetric    *string       `json:"SYSTEM_LOAD_CIRCUIT_BREAKER_METRIC" yaml:"system_load_circuit_breaker_metric,omitempty"`
-	MaxTxBandwidth                    *int64        `json:"MAX_TX_BANDWIDTH" yaml:"max_tx_bandwidth,omitempty"` // unit: bps
-	BandwidthProbeInterval            *int          `json:"BANDWIDTH_PROBE_INTERVAL" yaml:"bandwidth_probe_interval,omitempty"`
-	TapInterfaceRegex                 *string       `json:"TAP_INTERFACE_REGEX" yaml:"tap_interface_regex,omitempty"`
-	MaxEscapeSeconds                  *int          `json:"MAX_ESCAPE_SECONDS" yaml:"max_escape_seconds,omitempty"`
-	Mtu                               *int          `json:"MTU" yaml:"mtu,omitempty"`
-	OutputVlan                        *int          `json:"OUTPUT_VLAN" yaml:"output_vlan,omitempty"`
-	CollectorSocketType               *string       `json:"COLLECTOR_SOCKET_TYPE" yaml:"collector_socket_type,omitempty"`
-	CompressorSocketType              *string       `json:"COMPRESSOR_SOCKET_TYPE" yaml:"compressor_socket_type,omitempty"`
-	NpbSocketType                     *string       `json:"NPB_SOCKET_TYPE" yaml:"npb_socket_type,omitempty"`
-	NpbVlanMode                       *int          `json:"NPB_VLAN_MODE" yaml:"npb_vlan_mode,omitempty"`
-	CollectorEnabled                  *int          `json:"COLLECTOR_ENABLED" yaml:"collector_enabled,omitempty"`             // 0: disabled 1:enabled
-	VTapFlow1sEnabled                 *int          `json:"VTAP_FLOW_1S_ENABLED" yaml:"vtap_flow_1s_enabled,omitempty"`       // 0: disabled 1:enabled
-	L4LogTapTypes                     []int         `json:"L4_LOG_TAP_TYPES" yaml:"l4_log_tap_types,omitempty"`               // tap type info, separate by ","
-	L4LogIgnoreTapSides               []int         `json:"L4_LOG_IGNORE_TAP_SIDES" yaml:"l4_log_ignore_tap_sides,omitempty"` // separate by ","
-	L7LogIgnoreTapSides               []int         `json:"L7_LOG_IGNORE_TAP_SIDES" yaml:"l7_log_ignore_tap_sides,omitempty"` // separate by ","
-	NpbDedupEnabled                   *int          `json:"NPB_DEDUP_ENABLED" yaml:"npb_dedup_enabled,omitempty"`             // 0: disabled 1:enabled
-	PlatformEnabled                   *int          `json:"PLATFORM_ENABLED" yaml:"platform_enabled,omitempty"`               // 0: disabled 1:enabled
-	IfMacSource                       *int          `json:"IF_MAC_SOURCE" yaml:"if_mac_source,omitempty"`                     // 0: 接口MAC 1: 接口名称 2: 虚拟机MAC解析
-	VMXMLPath                         *string       `json:"VM_XML_PATH" yaml:"vm_xml_path,omitempty"`
-	ExtraNetnsRegex                   *string       `json:"EXTRA_NETNS_REGEX" yaml:"extra_netns_regex,omitempty"`
-	NatIPEnabled                      *int          `json:"NAT_IP_ENABLED" yaml:"nat_ip_enabled,omitempty"` // 0: disabled 1:enabled
-	CapturePacketSize                 *int          `json:"CAPTURE_PACKET_SIZE" yaml:"capture_packet_size,omitempty"`
-	InactiveServerPortEnabled         *int          `json:"INACTIVE_SERVER_PORT_ENABLED" yaml:"inactive_server_port_enabled,omitempty"` // 0: disabled 1:enabled
-	InactiveIPEnabled                 *int          `json:"INACTIVE_IP_ENABLED" yaml:"inactive_ip_enabled,omitempty"`                   // 0: disabled 1:enabled
-	LogThreshold                      *int          `json:"LOG_THRESHOLD" yaml:"log_threshold,omitempty"`
-	LogLevel                          *string       `json:"LOG_LEVEL" yaml:"log_level,omitempty"`
-	LogRetention                      *int          `json:"LOG_RETENTION" yaml:"log_retention,omitempty"`
-	HTTPLogProxyClient                *string       `json:"HTTP_LOG_PROXY_CLIENT" yaml:"http_log_proxy_client,omitempty"`
-	HTTPLogTraceID                    *string       `json:"HTTP_LOG_TRACE_ID" yaml:"http_log_trace_id,omitempty"`
-	L7LogPacketSize                   *int          `json:"L7_LOG_PACKET_SIZE" yaml:"l7_log_packet_size,omitempty"`
-	L4LogCollectNpsThreshold          *int          `json:"L4_LOG_COLLECT_NPS_THRESHOLD" yaml:"l4_log_collect_nps_threshold,omitempty"`
-	L7LogCollectNpsThreshold          *int          `json:"L7_LOG_COLLECT_NPS_THRESHOLD" yaml:"l7_log_collect_nps_threshold,omitempty"`
-	L7MetricsEnabled                  *int          `json:"L7_METRICS_ENABLED" yaml:"l7_metrics_enabled,omitempty"`         // 0: disabled 1:enabled
-	L7LogStoreTapTypes                []int         `json:"L7_LOG_STORE_TAP_TYPES" yaml:"l7_log_store_tap_types,omitempty"` // l7 log store tap types, separate by ","
-	CaptureSocketType                 *int          `json:"CAPTURE_SOCKET_TYPE" yaml:"capture_socket_type,omitempty"`
-	CaptureBpf                        *string       `json:"CAPTURE_BPF" yaml:"capture_bpf,omitempty"`
-	TapMode                           *int          `json:"TAP_MODE" yaml:"tap_mode,omitempty"`
-	ThreadThreshold                   *int          `json:"THREAD_THRESHOLD" yaml:"thread_threshold,omitempty"`
-	ProcessThreshold                  *int          `json:"PROCESS_THRESHOLD" yaml:"process_threshold,omitempty"`
-	Lcuuid                            *string       `json:"LCUUID" yaml:"-"`
-	NtpEnabled                        *int          `json:"NTP_ENABLED" yaml:"ntp_enabled,omitempty"`                         // 0: disabled 1:enabled
-	L4PerformanceEnabled              *int          `json:"L4_PERFORMANCE_ENABLED" yaml:"l4_performance_enabled,omitempty"`   // 0: disabled 1:enabled
-	PodClusterInternalIP              *int          `json:"POD_CLUSTER_INTERNAL_IP" yaml:"pod_cluster_internal_ip,omitempty"` // 0:  1:
-	Domains                           []string      `json:"DOMAINS" yaml:"domains,omitempty"`                                 // domains info, separate by ","
-	DecapType                         []int         `json:"DECAP_TYPE" yaml:"decap_type,omitempty"`                           // separate by ","
-	HTTPLogSpanID                     *string       `json:"HTTP_LOG_SPAN_ID" yaml:"http_log_span_id,omitempty"`
-	SysFreeMemoryLimit                *int          `json:"SYS_FREE_MEMORY_LIMIT" yaml:"sys_free_memory_limit,omitempty"` // unit: %
-	LogFileSize                       *int          `json:"LOG_FILE_SIZE" yaml:"log_file_size,omitempty"`                 // unit: MB
-	HTTPLogXRequestID                 *string       `json:"HTTP_LOG_X_REQUEST_ID" yaml:"http_log_x_request_id,omitempty"`
-	ExternalAgentHTTPProxyEnabled     *int          `json:"EXTERNAL_AGENT_HTTP_PROXY_ENABLED" yaml:"external_agent_http_proxy_enabled,omitempty"`
-	ExternalAgentHTTPProxyPort        *int          `json:"EXTERNAL_AGENT_HTTP_PROXY_PORT" yaml:"external_agent_http_proxy_port,omitempty"`
-	PrometheusHttpAPIAddresses        []string      `json:"PROMETHEUS_HTTP_API_ADDRESSES" yaml:"prometheus_http_api_addresses,omitempty"` // ip:port
-	AnalyzerPort                      *int          `json:"ANALYZER_PORT" yaml:"analyzer_port,omitempty"`
-	ProxyControllerPort               *int          `json:"PROXY_CONTROLLER_PORT" yaml:"proxy_controller_port,omitempty"`
-	ProxyControllerIP                 *string       `json:"PROXY_CONTROLLER_IP" yaml:"proxy_controller_ip,omitempty"`
-	AnalyzerIP                        *string       `json:"ANALYZER_IP" yaml:"analyzer_ip,omitempty"`
-	WasmPlugins                       []string      `json:"WASM_PLUGINS" yaml:"wasm_plugins,omitempty"`
-	SoPlugins                         []string      `json:"SO_PLUGINS" yaml:"so_plugins,omitempty"`
-	YamlConfig                        *StaticConfig `yaml:"static_config,omitempty"`
-}
-
-type TypeInfo struct {
-	ID   int    `json:"ID"`
-	Name string `json:"NAME"`
-}
-
-type DomainInfo struct {
-	ID   string `json:"ID"`
-	Name string `json:"NAME"`
-}
-
-type TapSideInfo struct {
-	ID   int    `json:"ID"`
-	Name string `json:"NAME"`
-}
-
-type VTapGroupConfigurationResponse struct {
-	MaxCollectPps                     *int           `json:"MAX_COLLECT_PPS"`
-	MaxNpbBps                         *int64         `json:"MAX_NPB_BPS"` // unit: bps
-	MaxCPUs                           *int           `json:"MAX_CPUS"`
-	MaxMemory                         *int           `json:"MAX_MEMORY"` // unit: M
-	SyncInterval                      *int           `json:"SYNC_INTERVAL"`
-	PlatformSyncInterval              *int           `json:"PLATFORM_SYNC_INTERVAL"`
-	StatsInterval                     *int           `json:"STATS_INTERVAL"`
-	RsyslogEnabled                    *int           `json:"RSYSLOG_ENABLED"` // 0: disabled 1:enabled
-	SystemLoadCircuitBreakerThreshold *float32       `json:"SYSTEM_LOAD_CIRCUIT_BREAKER_THRESHOLD"`
-	SystemLoadCircuitBreakerRecover   *float32       `json:"SYSTEM_LOAD_CIRCUIT_BREAKER_RECOVER"`
-	SystemLoadCircuitBreakerMetric    *string        `json:"SYSTEM_LOAD_CIRCUIT_BREAKER_METRIC"`
-	MaxTxBandwidth                    *int64         `json:"MAX_TX_BANDWIDTH"` // unit: bps
-	BandwidthProbeInterval            *int           `json:"BANDWIDTH_PROBE_INTERVAL"`
-	TapInterfaceRegex                 *string        `json:"TAP_INTERFACE_REGEX"`
-	MaxEscapeSeconds                  *int           `json:"MAX_ESCAPE_SECONDS"`
-	Mtu                               *int           `json:"MTU"`
-	OutputVlan                        *int           `json:"OUTPUT_VLAN"`
-	CollectorSocketType               *string        `json:"COLLECTOR_SOCKET_TYPE"`
-	CompressorSocketType              *string        `json:"COMPRESSOR_SOCKET_TYPE"`
-	NpbSocketType                     *string        `json:"NPB_SOCKET_TYPE"`
-	NpbVlanMode                       *int           `json:"NPB_VLAN_MODE"`
-	CollectorEnabled                  *int           `json:"COLLECTOR_ENABLED"`       // 0: disabled 1:enabled
-	VTapFlow1sEnabled                 *int           `json:"VTAP_FLOW_1S_ENABLED"`    // 0: disabled 1:enabled
-	L4LogTapTypes                     []*TypeInfo    `json:"L4_LOG_TAP_TYPES"`        // tap type info, separate by ","
-	L4LogIgnoreTapSides               []*TapSideInfo `json:"L4_LOG_IGNORE_TAP_SIDES"` // separate by ","
-	L7LogIgnoreTapSides               []*TapSideInfo `json:"L7_LOG_IGNORE_TAP_SIDES"` // separate by ","
-	NpbDedupEnabled                   *int           `json:"NPB_DEDUP_ENABLED"`       // 0: disabled 1:enabled
-	PlatformEnabled                   *int           `json:"PLATFORM_ENABLED"`        // 0: disabled 1:enabled
-	IfMacSource                       *int           `json:"IF_MAC_SOURCE"`           // 0: 接口MAC 1: 接口名称 2: 虚拟机MAC解析
-	VMXMLPath                         *string        `json:"VM_XML_PATH"`
-	ExtraNetnsRegex                   *string        `json:"EXTRA_NETNS_REGEX"`
-	NatIPEnabled                      *int           `json:"NAT_IP_ENABLED"` // 0: disabled 1:enabled
-	CapturePacketSize                 *int           `json:"CAPTURE_PACKET_SIZE"`
-	InactiveServerPortEnabled         *int           `json:"INACTIVE_SERVER_PORT_ENABLED"` // 0: disabled 1:enabled
-	InactiveIPEnabled                 *int           `json:"INACTIVE_IP_ENABLED"`          // 0: disabled 1:enabled
-	VTapGroupLcuuid                   *string        `json:"VTAP_GROUP_LCUUID"`
-	VTapGroupID                       *string        `json:"VTAP_GROUP_ID"`
-	VTapGroupName                     *string        `json:"VTAP_GROUP_NAME"`
-	LogThreshold                      *int           `json:"LOG_THRESHOLD"`
-	LogLevel                          *string        `json:"LOG_LEVEL"`
-	LogRetention                      *int           `json:"LOG_RETENTION"`
-	HTTPLogProxyClient                *string        `json:"HTTP_LOG_PROXY_CLIENT"`
-	HTTPLogTraceID                    *string        `json:"HTTP_LOG_TRACE_ID"`
-	L7LogPacketSize                   *int           `json:"L7_LOG_PACKET_SIZE"`
-	L4LogCollectNpsThreshold          *int           `json:"L4_LOG_COLLECT_NPS_THRESHOLD"`
-	L7LogCollectNpsThreshold          *int           `json:"L7_LOG_COLLECT_NPS_THRESHOLD"`
-	L7MetricsEnabled                  *int           `json:"L7_METRICS_ENABLED"`     // 0: disabled 1:enabled
-	L7LogStoreTapTypes                []*TypeInfo    `json:"L7_LOG_STORE_TAP_TYPES"` // l7 log store tap types, separate by ","
-	CaptureSocketType                 *int           `json:"CAPTURE_SOCKET_TYPE"`
-	CaptureBpf                        *string        `json:"CAPTURE_BPF"`
-	TapMode                           *int           `json:"TAP_MODE"`
-	ThreadThreshold                   *int           `json:"THREAD_THRESHOLD"`
-	ProcessThreshold                  *int           `json:"PROCESS_THRESHOLD"`
-	Lcuuid                            *string        `json:"LCUUID"`
-	NtpEnabled                        *int           `json:"NTP_ENABLED"`             // 0: disabled 1:enabled
-	L4PerformanceEnabled              *int           `json:"L4_PERFORMANCE_ENABLED"`  // 0: disabled 1:enabled
-	PodClusterInternalIP              *int           `json:"POD_CLUSTER_INTERNAL_IP"` // 0:  1:
-	Domains                           []*DomainInfo  `json:"DOMAINS"`                 // domains info, separate by ","
-	DecapType                         []*TypeInfo    `json:"DECAP_TYPE"`              // separate by ","
-	HTTPLogSpanID                     *string        `json:"HTTP_LOG_SPAN_ID"`
-	SysFreeMemoryLimit                *int           `json:"SYS_FREE_MEMORY_LIMIT"` // unit: %
-	LogFileSize                       *int           `json:"LOG_FILE_SIZE"`         // unit: MB
-	HTTPLogXRequestID                 *string        `json:"HTTP_LOG_X_REQUEST_ID"`
-	ExternalAgentHTTPProxyEnabled     *int           `json:"EXTERNAL_AGENT_HTTP_PROXY_ENABLED"`
-	ExternalAgentHTTPProxyPort        *int           `json:"EXTERNAL_AGENT_HTTP_PROXY_PORT"`
-	PrometheusHttpAPIAddresses        *string        `json:"PROMETHEUS_HTTP_API_ADDRESSES"` // separate by ","
-	AnalyzerPort                      *int           `json:"ANALYZER_PORT"`
-	ProxyControllerPort               *int           `json:"PROXY_CONTROLLER_PORT"`
-	ProxyControllerIP                 *string        `json:"PROXY_CONTROLLER_IP"`
-	AnalyzerIP                        *string        `json:"ANALYZER_IP"`
-	WasmPlugins                       []string       `json:"WASM_PLUGINS"`
-	SoPlugins                         []string       `json:"SO_PLUGINS"`
-}
-
 type DetailedConfig struct {
-	RealConfig    *VTapGroupConfigurationResponse `json:"REAL_CONFIG"`
-	DefaultConfig *VTapGroupConfigurationResponse `json:"DEFAULT_CONFIG"`
+	RealConfig    *agent_config.VTapGroupConfigurationResponse `json:"REAL_CONFIG"`
+	DefaultConfig *agent_config.VTapGroupConfigurationResponse `json:"DEFAULT_CONFIG"`
 }
 
 type VTapInterface struct {
